read and write filo metadata file without file transfer
When using pass-through mode of SCR, we'd still like to read and write
the filo metadata file. For example, this enables one to later do a
normal Filo_Fetch of a checkpoint that was written directly to the file
system rather than being transferred there through a Filo_flush. It may
be a better choice to define new Filo_read/write functions for this, but
this provides support for now.

During a flush, if the source and destination paths are the same for all
files on all ranks, no transfer is actually done, but filo will still
write its metadata.

During a fetch, if the target path is NULL, filo will read the metadata
file and return the list of source and destination file paths without
actually transferring the files.